verified with strict flags

+++ stash/str_bench.c
@@ -4,9 +4,9 @@
 #include "str.h"
 
 int main(void) {
-  long score = 0;
+  size_t score = 0;
   
-  for (int i = 0; i < 500000; ++i) {
+  for (size_t i = 0; i < 500000; ++i) {
     Str s = str_from_cstr("hello, world!");
     str_push_char(&s, 'a');
     str_push_cstr(&s, "xxx");
