+++ stash/str_bench.cpp
@@ -6,9 +6,9 @@
 using namespace std;
 
 int main(void) {
-  long score = 0;
+  size_t score = 0;
 
-  for (int i = 0; i < 500000; ++i) {
+  for (size_t i = 0; i < 500000; ++i) {
     string s("hello, world!");
     s += 'a';
     s += "xxx";
